+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Add `ContainerCreateOptsBuilder::envmerge`
 * Add `ContainerCreateOptsBuilder::groups`
 * Add `ContainerCreateOptsBuilder::health_check_on_failure_action`
+* Add `ContainerCreateOptsBuilder::hostusers`
 
 # 0.7.0
 * Make futures returned by methods Send + Sync